Use propagation for error handling
